requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiohttp-client-cache==0.11.1
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
arch==7.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
Brotli==1.1.0
cachetools==5.5.2
cattrs==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
courlan==1.3.2
cssselect==1.3.0
curl_adapter==1.1.0
curl_cffi==0.11.1
cycler==0.12.1
Cython==3.1.1
databricks-sdk==0.52.0
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
deepdiff==8.5.0
defusedxml==0.8.0rc2
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
et_xmlfile==2.0.0
exchange_calendars==4.10.1
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
finvizfinance==1.1.0
Flask==3.1.0
folium==0.19.6
fonttools==4.57.0
formulaic==1.1.1
fqdn==1.5.1
freezegun==1.5.2
frozendict==2.4.6
frozenlist==1.6.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.2
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
inscriptis==2.6.0
interface-meta==1.3.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
jusText==3.0.2
kagglehub==0.3.12
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
langdetect==1.0.9
lightgbm==4.6.0
linearmodels==6.1
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
missingno==0.5.2
mistune==3.1.3
mlflow==2.22.0
mlflow-skinny==2.22.0
monotonic==1.6
more-itertools==10.7.0
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.37.1
Nasdaq-Data-Link==1.0.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
notebook==7.4.1
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.5
openbb==4.4.4
openbb-alpha-vantage==1.4.1
openbb-benzinga==1.4.1
openbb-biztoc==1.4.2
openbb-bls==1.1.2
openbb-cboe==1.4.1
openbb-cftc==1.1.1
openbb-charting==2.3.4
openbb-commodity==1.3.1
openbb-core==1.4.7
openbb-crypto==1.4.1
openbb-currency==1.4.1
openbb-deribit==1.0.1
openbb-derivatives==1.4.1
openbb-ecb==1.4.2
openbb-econdb==1.3.1
openbb-econometrics==1.5.3
openbb-economy==1.4.2
openbb-equity==1.4.1
openbb-etf==1.4.1
openbb-federal-reserve==1.4.3
openbb-finra==1.4.1
openbb-finviz==1.3.1
openbb-fixedincome==1.4.3
openbb-fmp==1.4.2
openbb-fred==1.4.4
openbb-government-us==1.4.1
openbb-imf==1.1.1
openbb-index==1.4.1
openbb-intrinio==1.4.1
openbb-multpl==1.1.1
openbb-nasdaq==1.4.1
openbb-news==1.4.1
openbb-oecd==1.4.1
openbb-platform-api==1.1.10
openbb-polygon==1.4.1
openbb-quantitative==1.4.3
openbb-regulators==1.4.2
openbb-sec==1.4.3
openbb-seeking-alpha==1.4.1
openbb-stockgrid==1.4.1
openbb-technical==1.4.3
openbb-tiingo==1.4.1
openbb-tmx==1.3.2
openbb-tradier==1.3.1
openbb-tradingeconomics==1.4.1
openbb-us-eia==1.1.1
openbb-wsj==1.4.1
openbb-yfinance==1.4.6
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
optuna==4.3.0
orderly-set==5.4.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-ta-openbb==0.4.21
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==5.24.1
pluggy==1.6.0
posthog==3.25.0
praw==7.8.1
prawcore==2.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycurl==7.45.6
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
pyhdfe==0.2.0
PyJWT==2.10.1
pyluach==2.2.0
pyparsing==3.2.3
pytest==8.4.0
pytest-freezegun==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
random-user-agent==1.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9.1
ruff==0.7.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.3.0
setuptools-scm==8.3.1
sgmllib3k==1.0.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.45.1
streamlit_folium==0.25.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.52.4
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.4.0
uuid7==0.1.0
uvicorn==0.34.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.0
xlrd==2.0.1
xmltodict==0.13.0
xyzservices==2025.4.0
yarl==1.20.0
yfinance==0.2.58
zipp==3.21.0
